+++ e2e/tests/evm/timelock_inspection.go
@@ -14,8 +14,6 @@ import (
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/stretchr/testify/suite"
 
-	chainsel "github.com/smartcontractkit/chain-selectors"
-
 	e2e "github.com/smartcontractkit/mcms/e2e/tests"
 	testutils "github.com/smartcontractkit/mcms/e2e/utils"
 	"github.com/smartcontractkit/mcms/sdk/evm"
@@ -326,29 +324,3 @@ func (s *TimelockInspectionTestSuite) TestGetMinDelay() {
 	s.Require().EqualValues(0, delay)
 }
 
-func (s *TimelockInspectionTestSuite) TestUpdateDelay() {
-	ctx := context.Background()
-
-	timelockContract := testutils.DeployTimelockContract(&s.Suite, s.ClientA, s.auth, s.publicKey.String())
-	addr := timelockContract.Address().Hex()
-
-	configurer := evm.NewTimelockConfigurer(s.ClientA, s.auth)
-
-	delay, err := configurer.GetMinDelay(ctx, addr)
-	s.Require().NoError(err, "Failed to get initial min delay")
-	s.Require().EqualValues(0, delay)
-
-	newDelay := uint64(120)
-	result, err := configurer.UpdateDelay(ctx, addr, newDelay)
-	s.Require().NoError(err, "Failed to update delay")
-	s.Require().NotEmpty(result.Hash, "Transaction hash should not be empty")
-	s.Require().Equal(chainsel.FamilyEVM, result.ChainFamily, "Chain family should be EVM")
-
-	receipt, err := testutils.WaitMinedWithTxHash(ctx, s.ClientA, common.HexToHash(result.Hash))
-	s.Require().NoError(err, "Failed to wait for transaction to be mined")
-	s.Require().Equal(types.ReceiptStatusSuccessful, receipt.Status, "Transaction was not successful")
-
-	delay, err = configurer.GetMinDelay(ctx, addr)
-	s.Require().NoError(err, "Failed to get updated min delay")
-	s.Require().Equal(newDelay, delay, "Delay should match the updated value")
-}
