+++ e2e/tests/solana/timelock_inspection.go
@@ -8,8 +8,6 @@ import (
 
 	"github.com/gagliardetto/solana-go"
 
-	chainsel "github.com/smartcontractkit/chain-selectors"
-
 	"github.com/smartcontractkit/chainlink-ccip/chains/solana/gobindings/v0_1_1/timelock"
 	timelockutils "github.com/smartcontractkit/chainlink-ccip/chains/solana/utils/timelock"
 
@@ -174,32 +172,6 @@ func (s *TestSuite) TestGetMinDelay() {
 
 	s.Require().NoError(err, "Failed to query min delay")
 	s.Require().Equal(delay, uint64(minDelay.Seconds()), "Min delay should match the configured value")
-}
-
-func (s *TestSuite) TestUpdateDelay() {
-	ctx := context.Background()
-	initialDelay := 1 * time.Second
-	s.SetupTimelock(testPDASeedTimelockUpdateDelay, initialDelay)
-
-	admin, err := solana.PrivateKeyFromBase58(privateKey)
-	s.Require().NoError(err)
-
-	timelockAddr := solanasdk.ContractAddress(s.TimelockProgramID, testPDASeedTimelockUpdateDelay)
-	configurer := solanasdk.NewTimelockConfigurer(s.SolanaClient, admin)
-
-	delay, err := configurer.GetMinDelay(ctx, timelockAddr)
-	s.Require().NoError(err, "Failed to get initial min delay")
-	s.Require().Equal(uint64(initialDelay.Seconds()), delay, "Initial delay should match configured value")
-
-	newDelay := uint64(120)
-	result, err := configurer.UpdateDelay(ctx, timelockAddr, newDelay)
-	s.Require().NoError(err, "Failed to update delay")
-	s.Require().NotEmpty(result.Hash, "Transaction hash should not be empty")
-	s.Require().Equal(chainsel.FamilySolana, result.ChainFamily, "Chain family should be Solana")
-
-	delay, err = configurer.GetMinDelay(ctx, timelockAddr)
-	s.Require().NoError(err, "Failed to get updated min delay")
-	s.Require().Equal(newDelay, delay, "Delay should match the updated value")
 }
 
 func (s *TestSuite) createOperation(timelockID [32]byte) timelockutils.Operation {
