+++ sdk/timelock_configurer.go
@@ -8,6 +8,5 @@ import (
 
 // TimelockConfigurer is an interface for configuring timelock contract parameters.
 type TimelockConfigurer interface {
-	TimelockInspector
 	UpdateDelay(ctx context.Context, timelockAddress string, newDelay uint64) (types.TransactionResult, error)
 }
